fix(grid-filtering): Push the filtering menu below the grid header
Because of the way the filtering menu is currently implemented as part
of the header, it's visibility is affected by the z-index of the header
itself. So if a column header has a higher z-index than the next
sibling, that sibling will have its filtering ui covered by the previous
header. To avoid that I am pushing the filtering menu below the header
so it doesn't get covered by the headers.

Closes #1078