fix(igxTransaction): return correct aggregated value, #566
If we have state for a record and have pending state for the same record we
should merge the pending value over the state value and result over the
record ref. Up to this fix we have merged pending over record ref, state over
record ref and then both result one over other - which is not correct.