fix(igxOverlay): remove AnimationPlayer.init() call, #2867
Wright after creation of animationPlayer we were calling its init method. This
forces the player to start and then to pause. As a result on IE and on Edge
the animating element was flickering. Calling of init is not necessary as
calling play calls same internal methods as init does.