fix(igxCalendar): add trackBy for months and years, #3595
If calendar is shown in overlay, the overlay saves a reference
to calendar elementRef. Month and Year templates add click event
via host listener. However, both templates redraw themselves.
As a result of this elements, where the click event was added,
are lost. As a result click does not fire. To fix this we added
trackBy in ngFor both for Month and Year templates.