fix(igxToggle): add exclude position target, #3810
excludePositionTarget optional internal property added to overlay settings. If
set to true it will exclude this target from outside click, e.g. if set to true
and click is on the target overlay will not close.

fix for toggle, open and close methods of igxToggle. Now if there is close
animation open will start open animation and vice versa.