fix(grids): set z-index of row drag ghost to 1, #5748
This will make the z-index of the row drag ghost equal to the one of the grid.
This will not fix application scenario where grid is put in stacking container
with higher z-index. To fix this drag and drop directive should be able to use
outlet where drag ghost should be put in the DOM tree instead as a child
of the body.