fix(date-picker): updae ngModel on outside click, #6471
If we drop down the calendar when input has no value,
onBlur gets called but we do not call onTouchCallback
which is correct. However if after we click outside the
input the model is never validated. This is why we
should validate model, in this case, and in onClosing event.